% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerateblood.R
\docType{data}
\name{enumerateblood}
\alias{enumerateblood}
\title{Enumerate cell types in blood from gene expression.}
\format{A model object of class cv.glmnet (implemented in the glmnet R package)}
\source{
\url{add citation}
}
\usage{
enumerateblood
}
\description{
A multi-response Gaussian model capable of accurately estimating
the composition of blood samples from their gene expression profiles. Fit on
Affymetrix Gene ST gene expression profiles using the glmnet R package.
}
\keyword{datasets}

